rag_service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.6
python-dotenv==1.0.1
httpx==0.28.1

langchain==0.3.19
langchain-openai==0.3.6
langchain-qdrant==0.2.0
qdrant-client==1.13.2

SQLAlchemy==2.0.38
PyMySQL==1.1.1
tiktoken==0.8.0
